add/refactor utility functions for rollback availablity
