+++ src/olympia/addons/tests/test_utils_.py
@@ -5,8 +5,7 @@ from django.forms import ValidationError
 import pytest
 from freezegun import freeze_time
 
-from olympia import amo
-from olympia.amo.tests import TestCase, addon_factory, user_factory, version_factory
+from olympia.amo.tests import TestCase, addon_factory, user_factory
 from olympia.users.models import Group, GroupUser
 
 from ..utils import (
@@ -14,7 +13,6 @@ from ..utils import (
     DeleteTokenSigner,
     get_addon_recommendations,
     get_filtered_fallbacks,
-    validate_version_number_is_gt_latest_signed_listed_version,
     verify_mozilla_trademark,
 )
 
@@ -106,65 +104,3 @@ def test_delete_token_signer(frozen_time=None):
     # but not after 60 seconds
     frozen_time.tick(timedelta(seconds=2))
     assert not signer.validate(token, addon_id)
-
-
-@pytest.mark.django_db
-def test_validate_version_number_is_gt_latest_signed_listed_version():
-    addon = addon_factory(version_kw={'version': '123.0'}, file_kw={'is_signed': True})
-    # add an unlisted version, which should be ignored.
-    latest_unlisted = version_factory(
-        addon=addon,
-        version='124',
-        channel=amo.CHANNEL_UNLISTED,
-        file_kw={'is_signed': True},
-    )
-    # Version number is greater, but doesn't matter, because the check is listed only.
-    assert latest_unlisted.version > addon.current_version.version
-
-    # version number isn't greater (its the same).
-    assert validate_version_number_is_gt_latest_signed_listed_version(addon, '123') == (
-        'Version 123 must be greater than the previous approved version 123.0.'
-    )
-    # version number is less than the current listed version.
-    assert validate_version_number_is_gt_latest_signed_listed_version(
-        addon, '122.9'
-    ) == ('Version 122.9 must be greater than the previous approved version 123.0.')
-    # version number is greater, so no error message.
-    assert not validate_version_number_is_gt_latest_signed_listed_version(
-        addon, '123.1'
-    )
-
-    addon.current_version.file.update(is_signed=False)
-    # Same as current but check only applies to signed versions, so no error message.
-    assert not validate_version_number_is_gt_latest_signed_listed_version(addon, '123')
-
-    # Set up the scenario when a newer version has been signed, but then disabled
-    addon.current_version.file.update(is_signed=True)
-    disabled = version_factory(
-        addon=addon,
-        version='123.5',
-        file_kw={'is_signed': True, 'status': amo.STATUS_DISABLED},
-    )
-    addon.reload()
-    assert validate_version_number_is_gt_latest_signed_listed_version(
-        addon, '123.1'
-    ) == ('Version 123.1 must be greater than the previous approved version 123.5.')
-
-    disabled.delete()
-    # Shouldn't make a difference even if it's deleted - it was still signed.
-    assert validate_version_number_is_gt_latest_signed_listed_version(
-        addon, '123.1'
-    ) == ('Version 123.1 must be greater than the previous approved version 123.5.')
-
-    # Also check the edge case when addon is None
-    assert not validate_version_number_is_gt_latest_signed_listed_version(None, '123')
-
-
-@pytest.mark.django_db
-def test_validate_version_number_is_gt_latest_signed_listed_version_not_langpack():
-    addon = addon_factory(version_kw={'version': '123.0'}, file_kw={'is_signed': True})
-    assert validate_version_number_is_gt_latest_signed_listed_version(addon, '122') == (
-        'Version 122 must be greater than the previous approved version 123.0.'
-    )
-    addon.update(type=amo.ADDON_LPAPP)
-    assert not validate_version_number_is_gt_latest_signed_listed_version(addon, '122')
