+++ src/olympia/devhub/forms.py
@@ -37,7 +37,6 @@ from olympia.addons.utils import (
     fetch_translations_from_addon,
     get_translation_differences,
     remove_icons,
-    validate_version_number_is_gt_latest_signed_listed_version,
     verify_mozilla_trademark,
 )
 from olympia.amo.fields import HttpHttpsOnlyURLField, ReCaptchaField
@@ -71,6 +70,9 @@ from olympia.versions.models import (
     ApplicationsVersions,
     License,
     Version,
+)
+from olympia.versions.utils import (
+    validate_version_number_is_gt_latest_signed_listed_version,
 )
