+++ src/olympia/versions/tests/test_utils.py
@@ -14,7 +14,7 @@ from freezegun import freeze_time
 from PIL import Image, ImageChops
 
 from olympia import amo
-from olympia.amo.tests import root_storage
+from olympia.amo.tests import addon_factory, root_storage, version_factory
 from olympia.zadmin.models import set_config
 
 from ..utils import (
@@ -23,6 +23,7 @@ from ..utils import (
     get_review_due_date,
     get_staggered_review_due_date_generator,
     process_color_value,
+    validate_version_number_is_gt_latest_signed_listed_version,
     write_svg_to_png,
 )
 
@@ -307,3 +308,65 @@ def test_get_staggered_review_due_date_generator_garbage_config(log_mock):
         'extra-review-target-per-day',
         'lolweird',
     )
+
+
+@pytest.mark.django_db
+def test_validate_version_number_is_gt_latest_signed_listed_version():
+    addon = addon_factory(version_kw={'version': '123.0'}, file_kw={'is_signed': True})
+    # add an unlisted version, which should be ignored.
+    latest_unlisted = version_factory(
+        addon=addon,
+        version='124',
+        channel=amo.CHANNEL_UNLISTED,
+        file_kw={'is_signed': True},
+    )
+    # Version number is greater, but doesn't matter, because the check is listed only.
+    assert latest_unlisted.version > addon.current_version.version
+
+    # version number isn't greater (its the same).
+    assert validate_version_number_is_gt_latest_signed_listed_version(addon, '123') == (
+        'Version 123 must be greater than the previous approved version 123.0.'
+    )
+    # version number is less than the current listed version.
+    assert validate_version_number_is_gt_latest_signed_listed_version(
+        addon, '122.9'
+    ) == ('Version 122.9 must be greater than the previous approved version 123.0.')
+    # version number is greater, so no error message.
+    assert not validate_version_number_is_gt_latest_signed_listed_version(
+        addon, '123.1'
+    )
+
+    addon.current_version.file.update(is_signed=False)
+    # Same as current but check only applies to signed versions, so no error message.
+    assert not validate_version_number_is_gt_latest_signed_listed_version(addon, '123')
+
+    # Set up the scenario when a newer version has been signed, but then disabled
+    addon.current_version.file.update(is_signed=True)
+    disabled = version_factory(
+        addon=addon,
+        version='123.5',
+        file_kw={'is_signed': True, 'status': amo.STATUS_DISABLED},
+    )
+    addon.reload()
+    assert validate_version_number_is_gt_latest_signed_listed_version(
+        addon, '123.1'
+    ) == ('Version 123.1 must be greater than the previous approved version 123.5.')
+
+    disabled.delete()
+    # Shouldn't make a difference even if it's deleted - it was still signed.
+    assert validate_version_number_is_gt_latest_signed_listed_version(
+        addon, '123.1'
+    ) == ('Version 123.1 must be greater than the previous approved version 123.5.')
+
+    # Also check the edge case when addon is None
+    assert not validate_version_number_is_gt_latest_signed_listed_version(None, '123')
+
+
+@pytest.mark.django_db
+def test_validate_version_number_is_gt_latest_signed_listed_version_not_langpack():
+    addon = addon_factory(version_kw={'version': '123.0'}, file_kw={'is_signed': True})
+    assert validate_version_number_is_gt_latest_signed_listed_version(addon, '122') == (
+        'Version 122 must be greater than the previous approved version 123.0.'
+    )
+    addon.update(type=amo.ADDON_LPAPP)
+    assert not validate_version_number_is_gt_latest_signed_listed_version(addon, '122')
